non-standard topic names decompressed back to original topic name.
An extra string flag (no-basename) in the encoding.data string is passed into the encoding topic, which signifies to the decompressor that the topic name has no basename. It then uses the original topic name instead of appending image_raw. The encoding string is then returned to normal. This does not affect any of the standard topic names (ending with image_raw etc.)
Some unittests were extended to check the encoding string based on topic names.